AUDIO.JS CRASHES
Running the code for audio.js makes the browser crash, I think it has
something to do with the update() function towards the bottom of the
code. I commented out audio.init() in index.html to prevent this.

IF YOU HAVE ANY REASON TO NEED TO USE AUDIO CODE ASK ME FIRST